requirements.txt
aenum==3.1.15
aiohttp==3.9.1
aiosignal==1.3.1
annotated-types==0.6.0
async-timeout==4.0.3
attrs==23.2.0
blinker==1.7.0
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
Deprecated==1.2.14
Flask==3.0.1
frozenlist==1.4.1
future==0.18.3
idna==3.6
importlib-metadata==7.0.1
itsdangerous==2.1.2
Jinja2==3.1.3
line-bot-sdk==3.7.0
MarkupSafe==2.1.4
multidict==6.0.4
pydantic==2.5.3
pydantic_core==2.14.6
python-dateutil==2.8.2
requests==2.31.0
six==1.16.0
typing_extensions==4.9.0
urllib3==2.1.0
Werkzeug==3.0.1
wrapt==1.16.0
yarl==1.9.4
zipp==3.17.0

schedule~=1.2.1